+++ registrations/questions/misc_questions.py
@@ -3,62 +3,9 @@ from django.urls import reverse
 from django.template import loader
 
 from cdh.questions import questions
-from .models import Registration, ParticipantCategory, Involved, Receiver, Software
-from .progress import ProgressItemMixin
+from registrations.models import Registration, ParticipantCategory, Receiver, Software
 
-
-class RegistrationQuestionMixin(ProgressItemMixin):
-
-    show_progress = True
-    extra_form_kwargs = []
-
-    def __init__(self, *args, **kwargs):
-        self.blueprint = kwargs.pop("blueprint", None)
-        self.registration = kwargs.pop('registration', None)
-        self.view_kwargs = kwargs.pop('view_kwargs', None)
-        return super().__init__(*args, **kwargs)
-
-    def get_registration(self):
-        if not getattr(self, "registration"):
-            self.registration = self.blueprint.object
-        return self.registration
-
-    def get_blueprint(self):
-        return self.blueprint
-
-    def get_edit_url(self):
-
-        reverse_kwargs = {
-            'question': self.slug,
-            'question_pk': self.instance.pk,
-            'reg_pk': self.get_registration().pk,
-        }
-
-        if reverse_kwargs["question_pk"] is None:
-            reverse_kwargs.pop("question_pk")
-
-        return reverse(
-            'registrations:edit_question',
-            kwargs=reverse_kwargs,
-        )
-
-
-class PlaceholderQuestion(RegistrationQuestionMixin, questions.Question):
-    title = "Placeholder"
-    description = "Description of placeholder question"
-    is_editable = True
-
-    class Meta:
-        model = Registration
-        fields = []
-
-    def __init__(self, slug="placeholder", *args, **kwargs):
-        super().__init__(self)
-        self.slug = slug
-        self.disabled = True
-
-    def get_edit_url(self):
-        return False
+from .helpers import RegistrationQuestionMixin
 
 
 class NewRegistrationQuestion(
@@ -88,7 +35,10 @@ class NewRegistrationQuestion(
         return segments
 
 
-class FacultyQuestion(RegistrationQuestionMixin, questions.Question):
+class FacultyQuestion(
+        RegistrationQuestionMixin,
+        questions.Question
+):
 
     show_progress = True
     
@@ -213,221 +163,6 @@ class ConfirmInformationUseQuestion(RegistrationQuestionMixin, questions.Questio
         )
 
 
-class InvolvedPeopleQuestion(RegistrationQuestionMixin,
-                             questions.Question,):
-
-    class Meta:
-        model = Registration
-        fields = [
-            'involves_consent',
-            'involves_non_consent',
-            'involves_guardian_consent',
-            'involves_other_people',
-        ]
-
-    title = _("registrations:forms:involved_people_question_title")
-    description = _("registrations:forms:involved_people_question_description")
-    model = Meta.model
-    slug = "involved_people"
-    is_editable = True
-    show_progress = True
-
-    def get_segments(self):
-
-        return self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
-
-
-class QuestionViewArgumentsMixin():
-    """Use this mixin to receive arguments from the calling view"""
-
-    view_arguments = []
-
-    def __init__(self, *args, **kwargs):
-        for arg_name in self.view_arguments:
-            arg_value = kwargs.pop("arg_name", None)
-            setattr(self, arg_name, arg_value)
-        return super().__init__(*args, **kwargs)
-
-
-class NewInvolvedQuestion(
-        QuestionViewArgumentsMixin,
-        RegistrationQuestionMixin,
-        questions.Question,
-):
-    class Meta:
-        model = Involved
-        fields = [
-            'name',
-        ]
-
-    is_editable = True
-    slug = "new_involved"
-    title = _("registrations:forms:involved:new_title")
-    description = _("registrations:forms:involved:new_description")
-    model = Meta.model
-    view_arguments = ["group_type"]
-
-    def __init__(self, *args, **kwargs):
-        "Look for a group type in kwargs or alternatively given instance."
-
-        self.group_type = kwargs.pop("group_type", None)
-        if not self.group_type:
-            if "instance" in kwargs.keys():
-                self.group_type = kwargs["instance"].group_type
-        return super().__init__(*args, **kwargs)
-
-    def get_segments(self):
-        type_paragraph = questions.Segment(
-            type="paragraph",
-            paragraph=f"Type: {self.group_type}",
-        )
-
-        return [type_paragraph] + self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
-
-    def get_edit_url(self):
-
-        registration = self.get_registration()
-
-        reverse_kwargs = {
-            "reg_pk": registration.pk,
-            "question": self.slug,
-        }
-
-        if self.instance.pk:  # Edit existing model
-            reverse_kwargs['question_pk'] = self.instance.pk
-        else:  # Else, make sure we know what type to create
-            reverse_kwargs['group_type'] = self.instance.group_type
-
-        return reverse(
-            "registrations:edit_question",
-            kwargs=reverse_kwargs,
-        )
-    
-    def get_success_url(self):
-        return reverse(
-            "registrations:involved_manager",
-            kwargs={
-                "reg_pk": self.get_blueprint().object.pk,
-            })
-
-    def save(self, *args, **kwargs):
-        "Set a registration and group type on creation."
-        self.instance.registration = self.blueprint.object
-        return super().save(*args, **kwargs)
-    
-
-class PurposeQuestion(RegistrationQuestionMixin,
-                      questions.Question,
-                      ):
-
-    class Meta:
-        model = Involved
-        fields = [
-            "process_purpose",
-        ]
-
-    title = _("registrations:forms:purpose_question_title")
-    description = _("registrations:forms:purpose_question_description")
-    model = Meta.model
-    slug = "purpose"
-    is_editable = True
-    show_progress = True
-
-    def get_segments(self):
-
-        return self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
-
-
-class SpecialDetailsQuestion(
-        RegistrationQuestionMixin,
-        questions.Question,
-):
-
-    class Meta:
-        model = Involved
-        fields = [
-            "special_details",
-            "gave_explicit_permission",
-            "explicit_permission_details",
-            "grounds_for_exception",
-            "algemeen_belang_details",
-            "legal_grounds_details",
-        ]
-
-    title = _("registrations:special_details_title")
-    description = _("registrations:forms:special_details_description")
-    model = Meta.model
-    slug = "special_details"
-    is_editable = True
-    show_progress = True
-
-    def get_segments(self):
-
-        return self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
-
-
-class SensitiveDetailsQuestion(
-        RegistrationQuestionMixin,
-        questions.Question,
-):
-
-    class Meta:
-        model = Involved
-        fields = [
-            "provides_criminal_information",
-            "involves_children_under_15",
-            "other_sensitive_details",
-            "provides_no_sensitive_details",
-        ]
-
-    title = _("registrations:sensitive_details_title")
-    description = _("registrations:forms:sensitive_details_description")
-    model = Meta.model
-    slug = "sensitive_details"
-    is_editable = True
-    show_progress = True
-
-    def get_segments(self):
-
-        return self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
-
-
-class RegularDetailsQuestion(
-        RegistrationQuestionMixin,
-        questions.Question,
-):
-
-    class Meta:
-        model = Involved
-        fields = [
-            "regular_details",
-            "provides_ic_form",
-            "ic_form_details",
-            "extra_details",
-        ]
-
-    title = _("registrations:regular_details_title")
-    description = _("registrations:forms:regular_details_description")
-    model = Meta.model
-    slug = "regular_details"
-    is_editable = True
-    show_progress = True
-
-    def get_segments(self):
-
-        return self._fields_to_segments(
-            fields_list=self.Meta.fields,
-        )
 
 
 class RetentionQuestion(
@@ -793,21 +528,3 @@ class CategoryQuestion(RegistrationQuestionMixin, questions.Question):
         return super().save(*args, **kwargs)
 
 
-
-Q_LIST = [
-    NewRegistrationQuestion,
-    FacultyQuestion,
-    CategoryQuestion,
-    TraversalQuestion,
-    UsesInformationQuestion,
-    ConfirmInformationUseQuestion,
-    InvolvedPeopleQuestion,
-    RetentionQuestion,
-    NewInvolvedQuestion,
-    GoalQuestion,
-    PurposeQuestion,
-    NewSoftwareQuestion,
-    SoftwareQuestion,
-]
-
-QUESTIONS = {q.slug: q for q in Q_LIST}
